packages/bees/pyproject.toml
[project]
name = "bees"
version = "0.1.0"
description = "Bees — agent swarm orchestration framework"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sse-starlette>=2.0.0",
    "watchfiles>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["bees*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
